+++ src/dlc/UI/InGame/CivVAccess_ChatAccess.lua
@@ -58,6 +58,7 @@ include("CivVAccess_BaseMenuEditMode")
 include("CivVAccess_Help")
 include("CivVAccess_VolumeControl")
 include("CivVAccess_BeaconRange")
+include("CivVAccess_BeaconVolume")
 include("CivVAccess_Settings")
 
 local CHAT_HANDLER = "InGameChat"