+++ src/dlc/UI/InGame/CivVAccess_PlayerChangeAccess.lua
@@ -39,6 +39,7 @@ include("CivVAccess_BaseMenuEditMode")
 include("CivVAccess_Help")
 include("CivVAccess_VolumeControl")
 include("CivVAccess_BeaconRange")
+include("CivVAccess_BeaconVolume")
 include("CivVAccess_Settings")
 
 local priorShowHide = ShowHideHandler