+++ src/dlc/UI/InGame/LeaderHead/CivVAccess_DiscussionDialogAccess.lua
@@ -54,6 +54,7 @@ include("CivVAccess_LeaderDescription")
 include("CivVAccess_Help")
 include("CivVAccess_VolumeControl")
 include("CivVAccess_BeaconRange")
+include("CivVAccess_BeaconVolume")
 include("CivVAccess_Settings")
 
 local priorInput = InputHandler