+++ src/dlc/UI/InGame/WorldView/CivVAccess_SimpleDiploTradeAccess.lua
@@ -31,6 +31,7 @@ include("CivVAccess_TradeLogicAccess")
 include("CivVAccess_Help")
 include("CivVAccess_VolumeControl")
 include("CivVAccess_BeaconRange")
+include("CivVAccess_BeaconVolume")
 include("CivVAccess_Settings")
 
 local priorInput = InputHandler